:wqUpdate more TypeScript
