+++ package.json
@@ -8,6 +8,7 @@
   },
   "dependencies": {
     "@andrew-codes/gatsby-plugin-elasticlunr-search": "^1.0.4",
+    "@types/react-helmet": "^5.0.8",
     "array-flatten": "^2.1.1",
     "classnames": "^2.2.6",
     "debug": "^4.1.0",
