+++ src/web/components/AttributePeg.tsx
@@ -1,5 +1,4 @@
-/* @flow */
-import * as React from 'react'
+import React from 'react'
 
 export type Props = {
   hint: string
