+++ src/web/components/BackButton.module.css.d.ts
@@ -1,0 +1,3 @@
+export const home: string
+export const link: string
+export const icon: string
