+++ src/web/components/CommonHead.module.css
@@ -1,0 +1,80 @@
+@import 'sanitize.css/sanitize.css';
+@import 'responsive-modular-scale.css/responsive-modular-scale.css';
+@import 'hint.css/hint.css';
+
+/* Vars */
+@import '../styles/variables.css';
+
+/* Utils */
+@import '../styles/utils/gutter-padding.css';
+@import '../styles/utils/heading-style.css';
+@import '../styles/utils/section-gutter.css';
+@import '../styles/utils/has-container.css';
+
+/* Base */
+@import '../styles/base/base.css';
+
+/* Markdown */
+@import '../styles/markdown/a-em.css';
+@import '../styles/markdown/code.css';
+@import '../styles/markdown/headings.css';
+@import '../styles/markdown/table.css';
+@import '../styles/markdown/p.css';
+@import '../styles/markdown/prism.css';
+@import '../styles/markdown/ul.css';
+
+/* Components */
+@import '../styles/components/announcements-item.css';
+@import '../styles/components/announcements-list.css';
+@import '../styles/components/attribute-peg.css';
+@import '../styles/components/berry-sponsor.css';
+@import '../styles/components/body-area.css';
+@import '../styles/components/codefund-sponsor.css';
+@import '../styles/components/comments-section.css';
+@import '../styles/components/h2-section.css';
+@import '../styles/components/h3-section-list.css';
+@import '../styles/components/h3-section.css';
+@import '../styles/components/headline-ad.css'; /* TODO */
+@import '../styles/components/hint-mark.css';
+/* @import '../styles/components/home-button.css'; DONE */
+/* @import '../styles/components/intro-content.css'; DONE */
+@import '../styles/components/main-heading.css';
+@import '../styles/components/missing-message.css';
+@import '../styles/components/notice-box.css';
+@import '../styles/components/page-actions.css'; /* TODO */
+@import '../styles/components/pages-list.css';
+/* @import '../styles/components/pre-footer.css'; */
+@import '../styles/components/push-button.css'; /* TODO */
+@import '../styles/components/related-post-item.css';
+@import '../styles/components/related-post-list.css';
+@import '../styles/components/related-posts-area.css';
+/* @import '../styles/components/related-posts-callout.css'; DONE */
+/* @import '../styles/components/related-posts-group.css'; */
+@import '../styles/components/related-posts-section.css';
+/* @import '../styles/components/search-box.css'; DONE */
+/* @import '../styles/components/search-footer-container.css'; */
+/* @import '../styles/components/search-footer.css'; */
+@import '../styles/components/side-ad.css'; /* TODO */
+@import '../styles/components/site-header.css';
+/* @import '../styles/components/top-nav.css'; DONE */
+
+a,
+button,
+summary {
+  border-radius: 3px;
+  transition: background-color 100ms linear;
+}
+
+a:focus,
+button:focus,
+summary:focus {
+  outline: 0;
+  background-color: color-mod(var(--brand-b) alpha(6%));
+  animation: focus-in 300ms linear;
+}
+
+@keyframes focus-in {
+  0% {
+    background-color: color-mod(var(--brand-b) alpha(12%));
+  }
+}
